requirements.txt
streamlit
pillow
qrcode
matplotlib
pandas
numpy
